requirements.txt
django
django-tinymce
fontawesomefree
django-crispy-forms
crispy-bootstrap4
pillow
six